requirements-dev.txt
pytest>=8.0.0
pytest-cov>=4.1.0
pyyaml>=6.0.1
